fix: tests & refactor
