+++ utils/index.ts
@@ -11,3 +11,6 @@ export * from './transactionUtils';
 export * from './types';
 export * from './constants';
 export * from './txParser';
+export { formatLargeNumber } from '@/utils/format';
+export { formatAmount } from '@/utils/format';
+export { formatDenom } from '@/utils/format';
